Initial Web Scrape Code
This Code scrapes one url page of articles and extracts several variables from each article. There is a hitch in the code with extracting all the text from each article when exporting the output to a CSV file. The next step is scaling for the whole blog site, extracting from every page. Then we will need to scale for all urls.